pyproject.toml
[project]
name = "multiagent-framework-memory"
version = "1.0.0"
description = "Generic multi-agent framework on Databricks — with short-term + long-term memory via Lakebase"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.41.0",
    "databricks-openai[memory]>=0.13.0",
    "databricks-agents>=1.9.3",
    "databricks-sdk>=0.40.0",
    "mlflow>=3.10.0",
    "openai-agents>=0.4.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "databricks-langchain[memory]>=0.13.0",
    "psycopg[binary]>=3.1.0",
    "uuid-utils>=0.10.0",
    "mcp>=1.0.0",
    "pypdf>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent_server", "scripts", "mcp_servers"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
    "pytest>=9.0.2",
]

[project.scripts]
quickstart = "scripts.quickstart:main"
start-app = "scripts.start_app:main"
start-server = "agent_server.start_server:main"
agent-evaluate = "agent_server.evaluate_agent:evaluate"
discover-tools = "scripts.discover_tools:main"
preflight = "scripts.preflight:main"
